refactor with an extern types
