+++ package.json
@@ -15,10 +15,11 @@
         "#/*": "./src/*"
     },
     "dependencies": {
-        "@aws-sdk/client-s3": "3.1038.0",
+        "@aws-sdk/client-s3": "3.1039.0",
         "@js-temporal/polyfill": "0.5.1",
         "@prisma/adapter-pg": "7.8.0",
         "@prisma/client": "7.8.0",
+        "@triskcraft/api-types": "1.1.0",
         "argon2": "0.44.0",
         "cookie-parser": "1.4.7",
         "cors": "2.8.6",
